Images to items are added

+++ HRPortal.Web/Controllers/AdminApiController.cs
@@ -1,12 +1,16 @@
 ﻿using HRPortal.Business;
 using HRPortal.Model;
 using HRPortal.Model.ViewModels;
+using HRPortal.Web.Exceptions;
+using HRPortal.Web.Helpers;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using System.Web;
 using System.Web.Http;
 
 namespace HRPortal.Web.Controllers
@@ -63,11 +67,17 @@ namespace HRPortal.Web.Controllers
         }
 
         [HttpPost]
-        public async Task<IHttpActionResult> CreateNewProduct(TradeItem newItem)
+        public async Task<IHttpActionResult> CreateNewProduct()
         {
+            int catId;
+            decimal price;
+            if (!Int32.TryParse(HttpContext.Current.Request["CategoryId"], out catId) || !decimal.TryParse(HttpContext.Current.Request["Price"], out price)
+                || string.IsNullOrEmpty(HttpContext.Current.Request["Name"]) || string.IsNullOrEmpty(HttpContext.Current.Request["Description"]))
+                return Content(HttpStatusCode.BadRequest, "Model state invalid");
+
             try
             {
-                await tradeService.CreateProductAsync(newItem);
+                await tradeService.CreateProductAsync(GetNewsWithAttachmentsFromRequest(null, catId, price));
             }
             catch (Exception exp)
             {
@@ -79,7 +89,26 @@ namespace HRPortal.Web.Controllers
 
         public IHttpActionResult RemoveProduct(TradeItem item)
         {
-            tradeService.RemoveProduct(item);
+            var selectedItem = tradeService.GetProductById(item.Id);
+            if (selectedItem != null)
+            {
+                if (selectedItem.Attachments != null && selectedItem.Attachments.Any())
+                {
+                    foreach (var file in selectedItem.Attachments)
+                    {
+                        try
+                        {
+                            File.Delete(HttpContext.Current.Server.MapPath(file.AttachementPath));
+                        }
+                        catch (Exception exp)
+                        {
+                            logService.LogFatal(exp.Message);
+                            continue;
+                        }
+                    }
+                }
+                tradeService.RemoveProduct(item);
+            }
             return Ok();
         }
 
@@ -87,6 +116,52 @@ namespace HRPortal.Web.Controllers
         {
             tradeService.UpdateProduct(item);
             return Ok();
+        }
+
+        private TradeItem GetNewsWithAttachmentsFromRequest(int? id, int categoryId, decimal price)
+        {
+            List<TradeItemAttachment> attachedFiles = new List<TradeItemAttachment>();
+            if (!Directory.Exists(HttpContext.Current.Server.MapPath(AppConfig.UploadItemImagesTo)))
+            {
+                Directory.CreateDirectory(HttpContext.Current.Server.MapPath(AppConfig.UploadItemImagesTo));
+            }
+
+            for (int i = 0; i < HttpContext.Current.Request.Files.Count; i++)
+            {
+                if (i > 4)
+                    break;
+
+                var fileContent = HttpContext.Current.Request.Files[i];
+                if (!AppConfig.IsAttachmentAcceptable(fileContent.ContentType))
+                    throw new ImageException("File format is wrong!");
+
+                if (fileContent.ContentLength > 5242880)
+                    throw new ImageException("Image size is invalid");
+
+                if (fileContent != null && fileContent.ContentLength > 0)
+                {
+                    var fileName = Path.GetRandomFileName();
+                    fileName = Path.ChangeExtension(fileName, Path.GetExtension(fileContent.FileName));
+
+                    var path = Path.Combine(HttpContext.Current.Server.MapPath(AppConfig.UploadItemImagesTo), fileName);
+                    using (var fileStream = File.Create(path))
+                    {
+                        fileContent.InputStream.CopyTo(fileStream);
+                    }
+                    TradeItemAttachment attachedFile = new TradeItemAttachment { AttachementPath = $"{AppConfig.UploadItemImagesTo}{fileName}" };
+                    attachedFiles.Add(attachedFile);
+                }
+            }
+            return new TradeItem
+            {
+                Id = id.HasValue ? id.Value : 0,
+                CategoryId = categoryId,
+                Name = HttpContext.Current.Request["Name"],
+                Description = HttpContext.Current.Request["Description"],
+                Price = price,
+                CreateDate = DateTime.UtcNow,
+                Attachments = attachedFiles
+            };
         }
     }
 }
